Resturant_Project/requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
Django
djangorestframework
